requirements.txt
face_recognition==1.3.0
numpy==1.23.5
Pillow==9.3.0
python-telegram-bot==13.9